Properly handle issues getting the save directory. Closes #72

+++ src/main/java/the_fireplace/overlord/tools/Alliances.java
@@ -14,7 +14,7 @@ import java.util.UUID;
 public class Alliances implements Serializable {
     private static Alliances instance = null;
     private static final String dataFileName = "overlordalliances.dat";
-    private static final File saveDir = DimensionManager.getCurrentSaveRootDirectory();
+    private static File saveDir = DimensionManager.getCurrentSaveRootDirectory();
 
     private ArrayList<Alliance> alliances;
 
@@ -75,19 +75,21 @@ public class Alliances implements Serializable {
         readFromFile();
     }
 
-    @SuppressWarnings("TryWithIdenticalCatches")
     private static void readFromFile() {
+        if(saveDir == null)
+            saveDir = DimensionManager.getCurrentSaveRootDirectory();
+        if(saveDir == null) {
+            System.out.println("Error: Could not get save directory. Alliances will not load properly.");
+            instance = new Alliances();
+            return;
+        }
         File f = new File(saveDir, dataFileName);
         if (f.exists()) {
             try {
                 ObjectInputStream stream = new ObjectInputStream(new FileInputStream(f));
                 instance = (Alliances) stream.readObject();
                 stream.close();
-            } catch (IOException e) {
-                e.printStackTrace();
-                instance = new Alliances();
-                f.delete();
-            } catch (ClassNotFoundException e) {
+            } catch (IOException|ClassNotFoundException e) {
                 e.printStackTrace();
                 instance = new Alliances();
                 f.delete();
@@ -99,6 +101,10 @@ public class Alliances implements Serializable {
 
     private static void saveToFile() {
         try {
+            if(saveDir == null)
+                saveDir = DimensionManager.getCurrentSaveRootDirectory();
+            if(saveDir == null)
+                System.out.println("Error: Could not get save directory. Alliances will not save properly.");
             ObjectOutputStream out = new ObjectOutputStream(new FileOutputStream(new File(saveDir, dataFileName)));
             out.writeObject(instance);
             out.close();
